apipy/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
psycopg2-binary>=2.9.9
apscheduler>=3.10.4
akshare>=1.12.42
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0
httpx>=0.25.0